# openapi/openapi.yaml
openapi: 3.0.3
info:
  title: Allyourbase API
  description: |
    Backend-as-a-Service for PostgreSQL. Single binary. Auto-generated REST API
    with filtering, sorting, pagination, foreign key expansion, full-text search,
    batch operations, authentication, file storage, realtime events, and webhooks.
  version: 0.1.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  contact:
    name: Allyourbase
    url: https://allyourbase.io
servers:
  - url: http://localhost:8090
    description: Local development server

# Default: no auth required. Individual operations override with BearerAuth or AdminAuth.
security:
  - {}
...
paths:
...
  /api/admin/auth:
    post:
      tags: [Admin]
      summary: Admin login
      operationId: adminLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminLoginResponse"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Admin auth not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      operationId: authRegister
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid request (missing email/password)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded (10 req/min/IP)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      operationId: authRefresh
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Tokens refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/logout:
    post:
      tags: [Auth]
      summary: Logout (revoke refresh token)
      operationId: authLogout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "204":
          description: Logged out successfully
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    delete:
      tags: [Auth]
      summary: Delete own account
      description: Permanently deletes the authenticated user's account and all sessions.
      operationId: authDeleteAccount
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Account deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      tags: [Auth]
      summary: Get current user
      operationId: authMe
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/password-reset:
    post:
      tags: [Auth]
      summary: Request password reset email
      operationId: authPasswordReset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetRequest"
      responses:
        "204":
          description: Reset email sent (always returns 204 regardless of email existence)
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/password-reset/confirm:
    post:
      tags: [Auth]
      summary: Confirm password reset
      operationId: authPasswordResetConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PasswordResetConfirmRequest"
      responses:
        "204":
          description: Password reset successful
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/verify:
    post:
      tags: [Auth]
      summary: Verify email address
      operationId: authVerifyEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "204":
          description: Email verified
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/verify/resend:
    post:
      tags: [Auth]
      summary: Resend verification email
      operationId: authResendVerification
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Verification email sent
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/oauth/{provider}:
    get:
      tags: [Auth]
      summary: Start OAuth flow
      operationId: authOAuthRedirect
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github]
      responses:
        "200":
          description: Not used (endpoint redirects)
        "302":
          description: Redirect to OAuth provider
        "400":
          description: Provider not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/oauth/{provider}/callback:
    get:
      tags: [Auth]
      summary: OAuth callback
      operationId: authOAuthCallback
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [google, github]
        - name: code
          in: query
          schema:
            type: string
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Not used (endpoint redirects)
        "302":
          description: Redirect to app with tokens in hash fragment
        "400":
          description: Invalid OAuth callback
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/magic-link:
    post:
      tags: [Auth]
      summary: Request magic link email
      description: |
        Send a passwordless login link to the given email address. Always returns 200
        regardless of whether the email exists (prevents email enumeration). Returns 404
        if magic link feature is disabled in server config.
      operationId: authMagicLink
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkRequest"
      responses:
        "200":
          description: Magic link sent (always returned to prevent email enumeration)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Missing email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Magic link feature not enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/magic-link/confirm:
    post:
      tags: [Auth]
      summary: Confirm magic link token
      description: |
        Verify a magic link token and return auth tokens. Creates the user account
        if it doesn't exist (email is marked verified). Token is single-use.
      operationId: authMagicLinkConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkConfirmRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Magic link feature not enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/api-keys:
    get:
      tags: [Auth]
      summary: List own API keys
      description: List all API keys belonging to the authenticated user.
      operationId: authListApiKeys
      security:
        - BearerAuth: []
      responses:
        "200":
          description: User's API keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApiKey"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Auth]
      summary: Create an API key
      description: |
        Create a non-expiring API key for the authenticated user. Requires JWT
        authentication (not API key auth) to prevent key bootstrapping.
      operationId: authCreateApiKey
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreateApiKeyRequest"
      responses:
        "201":
          description: API key created (plaintext key shown once only)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiKeyCreateResponse"
        "400":
          description: Invalid scope or missing name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from /api/auth/login or /api/auth/register
    AdminAuth:
      type: http
      scheme: bearer
      description: Admin token from /api/admin/auth
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "collection not found: nonexistent"
        data:
          type: object
          description: Field-level validation details
          additionalProperties: true
...
    AuthRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: secretpassword
    AuthResponse:
      type: object
      required: [token, refreshToken, user]
      properties:
        token:
          type: string
          description: JWT access token (default 15 min expiry)
        refreshToken:
          type: string
          description: Refresh token (default 7 day expiry)
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, createdAt, updatedAt]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    PasswordResetConfirmRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          format: password
          minLength: 8
    VerifyEmailRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string
...
    AdminLoginRequest:
      type: object
      required: [password]
      properties:
        password:
          type: string
    AdminLoginResponse:
      type: object
      required: [token]
      properties:
        token:
          type: string
          description: Admin Bearer token (HMAC-derived, valid for server lifetime)
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    ApiKey:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        name:
          type: string
        keyPrefix:
          type: string
          description: "First 12 chars of key (e.g. ayb_a1b2c3d4)"
        scope:
          type: string
          enum: ["*", readonly, readwrite]
        allowedTables:
          type: array
          items:
            type: string
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        revokedAt:
          type: string
          format: date-time
          nullable: true
...
    UserCreateApiKeyRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        scope:
          type: string
          enum: ["*", readonly, readwrite]
          default: "*"
        allowedTables:
          type: array
          items:
            type: string
          description: "Empty = all tables"
...
    MagicLinkRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    MagicLinkConfirmRequest:
      type: object
      required: [token]
      properties:
        token:
          type: string